requirements.txt
streamlit
pandas
requests
python-telegram-bot==13.15
pyyaml          # basket
